+++ src/commands/add.js
