+++ src/commands/badbuilder.js
