+++ src/commands/deny.js
