+++ src/commands/friday.js
