+++ src/commands/see.js
