+++ src/deploy.js
@@ -1,7 +1,7 @@
 const fs = require('fs')
 const { REST } = require('@discordjs/rest')
 const { Routes } = require('discord-api-types/v9')
-const { clientId, guildId, token } = require('./config.json')
+const { clientId, guildId, token } = require('../config.json')
 
 const commands = []
 const commandFiles = fs
