+++ src/events/messageCreate.js
@@ -1,4 +1,4 @@
-const { prefix } = require('../config.json')
+const { prefix } = require('../../config.json')
 const Discord = require('discord.js')
 const cooldowns = new Discord.Collection()
