+++ src/slash/many.js
